bevy_animation::gltf_curves
Struct CubicKeyframeCurve

pub struct CubicKeyframeCurve<T> { /* private fields */ }

A keyframe-defined curve that uses cubic spline interpolation, backed by a contiguous buffer.

Implementations

impl<T> CubicKeyframeCurve<T>

pub fn new( times: impl IntoIterator<Item = f32>, values: impl IntoIterator<Item = T>, ) -> Result<Self, ChunkedUnevenCoreError>

Create a new CubicKeyframeCurve from keyframe times and their associated values. Because 3 values are needed to perform cubic interpolation, values must have triple the length of times — each consecutive triple a_k, v_k, b_k associated to time t_k consists of:

The in-tangent a_k for the sample at time t_k

The actual value v_k for the sample at time t_k

The out-tangent b_k for the sample at time t_k

For example, for a curve built from two keyframes, the inputs would have the following form:

times: [t_0, t_1]

values: [a_0, v_0, b_0, a_1, v_1, b_1]