bevy_animation
Module graph

The animation graph, which allows animations to be blended together.

Structs

AnimationGraph

A graph structure that describes how animation clips are to be blended together.

AnimationGraphAssetLoader

An AssetLoader that can load AnimationGraphs as assets.

AnimationGraphHandle

A Handle to the AnimationGraph to be used by the AnimationPlayer on the same entity.

AnimationGraphNode

An individual node within an animation graph.

SerializedAnimationGraph

A version of AnimationGraph suitable for serializing as an asset.

SerializedAnimationGraphNode

A version of AnimationGraphNode suitable for serializing as an asset.

ThreadedAnimationGraph

An acceleration structure for an animation graph that allows Bevy to evaluate it quickly.

ThreadedAnimationGraphs

Acceleration structures for animation graphs that allows Bevy to evaluate them quickly.

Enums

AnimationGraphLoadError

Various errors that can occur when serializing or deserializing animation graphs to and from RON, respectively.

AnimationNodeType

Animation node data specific to the type of node (clip, blend, or add).

SerializedAnimationClip

A version of Handle<AnimationClip> suitable for serializing as an asset.

SerializedAnimationNodeType

A version of AnimationNodeType suitable for serializing as part of a SerializedAnimationGraphNode asset.

Type Aliases

AnimationDiGraph

A type alias for the petgraph data structure that defines the animation graph.

AnimationMask

The type of an animation mask bitfield.

AnimationNodeIndex

The index of either an animation or blend node in the animation graph.