Crate bevy_animation

Animation for the game engine Bevy

Modules

prelude

Structs

ActiveAnimation

An animation that an AnimationPlayer is currently either playing or was playing, but is presently paused.

AnimationClip

A list of VariableCurves and the AnimationTargetIds to which they apply.

AnimationGraphEvaluator

Information needed during the traversal of the animation graph in advance_animations.

AnimationPlayer

Animation controls

AnimationPlugin

Adds animation support to an app

AnimationTarget

An entity that can be animated by an AnimationPlayer.

AnimationTargetId

A unique UUID for an animation target (e.g. bone in a skinned mesh).

VariableCurve

Describes how an attribute of a Transform or MorphWeights should be animated.

Enums

Interpolation

Interpolation method to use between keyframes.

Keyframes

List of keyframes for one of the attribute of a Transform.

RepeatAnimation

Repetition behavior of an animation.

Statics

ANIMATION_TARGET_NAMESPACE

The UUID namespace of animation targets (e.g. bones).

Functions

advance_animations

A system that advances the time for all playing animations.

animate_targets

A system that modifies animation targets (e.g. bones in a skinned mesh) according to the currently-playing animation.

Type Aliases

AnimationCurves

A mapping from AnimationTargetId (e.g. bone in a skinned mesh) to the animation curves.